src/main/java/sample/cafekiosk/spring/api/service/mail/MailService.java
package sample.cafekiosk.spring.api.service.mail;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import sample.cafekiosk.spring.client.mail.MailSendClient;
import sample.cafekiosk.spring.domain.history.mail.MailSendHistory;
import sample.cafekiosk.spring.domain.history.mail.MailSendHistoryRepository;

@Service
@RequiredArgsConstructor
public class MailService {

    private final MailSendClient mailSendClient;
    private final MailSendHistoryRepository mailSendHistoryRepository;

    public boolean sendMail(String fromEmail, String email, String title, String content) {
        boolean result = mailSendClient.sendMail(fromEmail, email, title, content);
        if (result) {
            mailSendHistoryRepository.save(
                    MailSendHistory.builder()
                            .fromEmail(fromEmail)
                            .toEmail(email)
                            .subject(title)
                            .content(content)
                            .build());
        }

        mailSendClient.a();
        mailSendClient.b();
        mailSendClient.c();
        return true;
    }
}
